requirements.txt
django-jazzmin==3.0.1
requests
Django==5.1.2
pillow==11.0.0
requests==2.32.3
psycopg2
psycopg2-binary
django-ckeditor
nbconvert
